pyproject.toml
[project]
name = "reflex_tap"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "ppspps824" , email = "papasim824@gmail.com"}
]
dependencies = [
    "reflex>=0.5.5",
    "pydantic==1.10.17",
    "python-dotenv>=1.0.1",
    "psycopg2>=2.9.9",
]
readme = "README.md"
requires-python = ">= 3.11"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.2.2",
    "pre-commit>=3.7.0",
    "ruff>=0.4.7",
]
